# tests/Swagger/mollie.yaml
openapi: "3.0.0"
info:
  title: "Mollie Shopware Plugin API"
  description: "This is a Swagger Template for all uses cases of the Mollie plugin.<br /><br />Instructions:<ol><li>Configure the correct server domain, where your Shopware API ist available.</li><li>Then use the authorize request for Shopware to retrieve an Access-Token. Use the default credentials or just enter your custom credentials.</li><li>Just copy the Access-Token and use it in the global authorize section above the request list.</li></ol><br />Congratulations, you can now use the API requests."
servers:
  - url: /
  - url: https://localhost
  - url: http://localhost
  - url: '{custom}'
    variables:
      custom:
        default: https://localhost
...
paths:
...
  /api/mollie/refund/order?number={number}&description={description}:
    get:
      tags:
        - "Refunds (Operational)"
      summary: "Full Refund"
      security:
        - AdminAPI: []
      parameters:
        - name: "number"
          in: "path"
          description: "Shopware order number"
          required: true
        - name: "description"
          in: "path"
          description: "(optional) custom description"
      responses:
        "200":
          description: "successful operation"
  /api/mollie/refund/order?number={number}&amount={amount}&description={description}:
    get:
      tags:
        - "Refunds (Operational)"
      summary: "Partial Refund"
      security:
        - AdminAPI: []
      parameters:
        - name: "number"
          in: "path"
          description: "Shopware order number"
          required: true
        - name: "amount"
          in: "path"
          description: "(optional) partial amount"
        - name: "description"
          in: "path"
          description: "(optional) custom description"
      responses:
        "200":
          description: "successful operation"
  /api/_action/mollie/refund-manager/data:
    post:
      tags:
        - "Refunds (Technical)"
      summary: "Get data from Refund Manager"
      security:
        - AdminAPI: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  default: ""
      responses:
        "200":
          description: "successful operation"
  /api/_action/mollie/refund:
    post:
      tags:
        - "Refunds (Technical)"
      summary: "Refund with Refund Manager"
      security:
        - AdminAPI: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  default: ""
                amount:
                  default: "0.00"
                description:
                  default: ""
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        default: ""
                      amount:
                        default: "0.0"
                      quantity:
                        default: "0"
                      resetStock:
                        default: "0"

      responses:
        "200":
          description: "successful operation"
...
components:
  securitySchemes:
    AdminAPI:
      type: apiKey
      in: header
      name: Authorization